fix: errors not propagating in some situations
BT-49
