+++ suggestions/utility/__init__.py
@@ -1,0 +1,1 @@
+from .error_wrapper import wrap_with_error_handler
